Unify `try_view_control_flow`s into one function
