+++ crates/circuit/src/instruction.rs
@@ -12,7 +12,6 @@
 
 use crate::circuit_data::CircuitData;
 use crate::operations::{OperationRef, Param};
-use crate::parameter::symbol_expr::Symbol;
 use ndarray::Array2;
 use num_complex::Complex64;
 use pyo3::prelude::*;
